jupyterhub: cleanup of unused options and files
