--- services/jupyterhub/secrets/jupyterhub.crt
@@ -1,17 +1,0 @@
------BEGIN CERTIFICATE-----
-MIICsDCCAhmgAwIBAgIJAMIWLzy67o6QMA0GCSqGSIb3DQEBBQUAMEUxCzAJBgNV
-BAYTAkFVMRMwEQYDVQQIEwpTb21lLVN0YXRlMSEwHwYDVQQKExhJbnRlcm5ldCBX
-aWRnaXRzIFB0eSBMdGQwHhcNMTgwMzA5MjMwNjM2WhcNMTkwMzA5MjMwNjM2WjBF
-MQswCQYDVQQGEwJBVTETMBEGA1UECBMKU29tZS1TdGF0ZTEhMB8GA1UEChMYSW50
-ZXJuZXQgV2lkZ2l0cyBQdHkgTHRkMIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKB
-gQDWSZxr687rGQK8qYYqI87tAWbZuyJiy1NvT8ZFZfvHTFMoYthEvyRReo+4Q/yV
-oSWRik/pR7L0Bevx3hAlyNtensDQBf5JRZobScWR9L+lxcpEldRrclpI71a1yLyn
-QrmUcXCISE6ttP55CQUgkNK28o47drmJYFmvhhJeztkQowIDAQABo4GnMIGkMB0G
-A1UdDgQWBBQIWZctIqcBUvcmy+r3fJV4scLc1DB1BgNVHSMEbjBsgBQIWZctIqcB
-Uvcmy+r3fJV4scLc1KFJpEcwRTELMAkGA1UEBhMCQVUxEzARBgNVBAgTClNvbWUt
-U3RhdGUxITAfBgNVBAoTGEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZIIJAMIWLzy6
-7o6QMAwGA1UdEwQFMAMBAf8wDQYJKoZIhvcNAQEFBQADgYEAzhJzh+6nCQ74NLbT
-rG73fkUIbCGQXRFgm1Aq9vRSIOIczFGLFo4XZiX1FuWQUa0UDPVqdmEwDN0HtTZY
-K7MYIr3tgSMFNgXzPOxcE74XjBqLRwI8p7XCoVkMjIAB/WgBlz4lb3mNvVCv6QYG
-CSOBwyJ9nX0XSMDMyPx3R28FnIk=
------END CERTIFICATE-----
