--- services/jupyterhub/secrets/jupyterhub.key
@@ -1,15 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIICXQIBAAKBgQDWSZxr687rGQK8qYYqI87tAWbZuyJiy1NvT8ZFZfvHTFMoYthE
-vyRReo+4Q/yVoSWRik/pR7L0Bevx3hAlyNtensDQBf5JRZobScWR9L+lxcpEldRr
-clpI71a1yLynQrmUcXCISE6ttP55CQUgkNK28o47drmJYFmvhhJeztkQowIDAQAB
-AoGAQcRh37WF0h3g+ypEZRaMYa02dci7avrgZp9jSNVeCvLf6ts7aQ4JBYDKL1Ge
-agYFU7/RetVolQBkQ91Vs2MMAFQRF5+BPNFTWY/ZqJcTExuwqLXuOSWuHRGzMg/j
-tqGIqfv26Lf9v7dDdnCyrEblK+L1KqFGhbzOMLhnlekKwrECQQD+fMXaXvoI2qMC
-wHE9CGM1dxUYLflGLEcXdAXBKVx9BpVm8PbMtZgPP92OTDOqIKOY4wA2QgIe73WY
-SS5eNDNFAkEA14+rkoaJrLJMYL1J1+urWhVgOr4n/4qNPJtxPnWA0paewzaUk6wg
-6TxoxWsY6w4v263bylDmfarpygsWhNi+xwJBAIIYiRC0Yv/pvVGnzXvU8NqhMNB2
-mCsucX1k25g+q8BigDfq0VAfOUf6bynSfMro+pkq2q3iGrFz2trp9mFEMD0CQQC2
-TtD5nIVlpGoKSyzfy3ZhOsDnDZoAeqqWvpRZEcHA2Rgu28Nu3V38RFYL2Vb6vUXm
-kjTX9/HoCJuvgA8sja9TAkBZC2PcEsz28zyiL00KoF9kS48yX84Z7+/RSU0ZjMf1
-I9MjAcyGhlN0gQdVCxhNGTlS0Rz1NsXFM/VMM0y/POBf
------END RSA PRIVATE KEY-----
